Rename betterdict submodule to dict
This makes this more consistent with the list submodule.
