+++ querylist/__init__.py
@@ -1,6 +1,6 @@
 """QueryList provides a simple way to filter lists of objects."""
 from querylist.list import QueryList
-from querylist.betterdict import BetterDict
+from querylist.dict import BetterDict
 
 __version__ = "0.2.0"
 __author__ = "Thomas Welfley"
